Moved old files and organised the necessary files
